# Conflicts: #	app/src/main/AndroidManifest.xml #	app/src/main/java/com/edu_app/model/NetworkUtility.java #	app/src/main/java/com/edu_app/view/teacher/person/activityContactService.java

+++ app/src/main/java/com/edu_app/model/teacher/practice/QuestionItem.java
@@ -74,7 +74,7 @@ public class QuestionItem implements Model, Serializable {
         return question.getScore();
     }
 
-    public void setScore(double score){
+    public void setScore(int score){
         question.setScore(score);
     }
 }
